Server/wsgi.py
#!/usr/bin/env python3
"""
WSGI entry point for production deployment with Gunicorn
"""
import os
import sys
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Add the current directory to Python path
sys.path.insert(0, os.path.dirname(__file__))

from app import app

if __name__ == "__main__":
    app.run()