Server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
google-generativeai==0.3.0
yfinance==0.2.18
requests==2.31.0
werkzeug==2.3.7
pyjwt==2.8.0
gunicorn==21.2.0